% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_log.R
\name{chi_log}
\alias{chi_log}
\title{Pearson's Goodness of Fit for a Logistic Regression Model}
\usage{
chi_log(model, data, name = "model")
}
\arguments{
\item{model}{The name of the object containing your logistic regression model.}

\item{data}{The name of your data frame.}

\item{name}{An optional string, labeling the output of the function.}
}
\description{
Use this function to easily calculate the chi-squared, degrees of freedom, and p-value for your model.
}
\examples{
chi_log(model_block1, demo, name = "Block 1")
chi_log(model_block2, demo, name = "Block 2")
}
