% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare Summary Results of Models}
\usage{
compare_models(..., labels = NULL)
}
\arguments{
\item{...}{list the names of all models you want to compare}

\item{labels}{an optional vector of names to label each model included in the output.}
}
\description{
Use this function to compare output from multiple models in one list of output.
}
\examples{
compare_models(mb1, mb2, labels = c("Block 1", "Block 2"))

}
