% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance_aic.R
\name{deviance_aic}
\alias{deviance_aic}
\title{Compare Deviance and AIC of models}
\usage{
deviance_aic(...)
}
\arguments{
\item{...}{The names of the models you wish to compare}
}
\description{
This function pulls the null deviance, residual deviance, and AIC from multiple models and puts them together in an easy to read output.
}
\examples{
deviance_aic(mb1, mb2, mb3)

}
