% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fences.R
\name{fences}
\alias{fences}
\title{Calculate Upper and Lower Fences}
\usage{
fences(x)
}
\arguments{
\item{x}{A continuous variable}
}
\description{
This function inputs a continous variable and returns the upper and lower fences.
}
\examples{
fences(demo$variable4)

}
