% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or.R
\name{or}
\alias{or}
\title{Odds ratios and Confidence Intervals from a Logistic Regression Model}
\usage{
or(x)
}
\arguments{
\item{x}{Name of the object containing the logistic regression model.}
}
\description{
This function takes a logistic regression model and returns a data frame that contains odds ratios and confidence intervals.
}
\examples{
or(model_block1)
}
