% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_percent.R
\name{predict_percent}
\alias{predict_percent}
\title{Accuracy for a Logistic Regression Model}
\usage{
predict_percent(model, dependent, name = "your model")
}
\arguments{
\item{model}{The name of the object containing your logistic regression model.}

\item{dependent}{the name of your dependent variable. Must use full syntax specifying the data frame the variable comes from. This will look like 'dataframe$variable'.}

\item{name}{An optional string, labeling the output of the function.}
}
\description{
This function calcualtes the accuracy of a logistic regression model based on the predicted probabilities.
}
\examples{
predict_percent(model_block1, demo$dependentvariable, name = "Block 1")
}
